+++ Exercism/c/beer-song/src/beer_song.h
@@ -1,0 +1,8 @@
+#ifndef BEER_SONG_H
+#define BEER_SONG_H
+
+void verse(char*, int);
+
+void sing(char*, int, int);
+
+#endif