+++ Exercism/c/binary-search/src/binary_search.h
@@ -1,0 +1,6 @@
+#ifndef BINARY_SEARCH_H
+#define BINARY_SEARCH_H
+
+int* binary_search(int x, int* arr, int len);
+
+#endif