+++ Exercism/c/difference-of-squares/src/difference_of_squares.h
@@ -1,0 +1,10 @@
+#ifndef DIFFERENCE_OF_SQUARES_H
+#define DIFFERENCE_OF_SQUARES_H
+
+long square_of_sum(int n);
+
+long sum_of_squares(int n);
+
+long difference_of_squares(int n);
+
+#endif