+++ Exercism/c/meetup/src/meetup.h
@@ -1,0 +1,6 @@
+#ifndef MEETUP_H
+#define MEETUP_H
+
+int meetupDayOfMonth(int year, int month, char* week, char* dayOfWeek);
+
+#endif