+++ Exercism/c/nth-prime/src/nth_prime.h
@@ -1,0 +1,6 @@
+#ifndef NTH_PRIME_H
+#define NTH_PRIME_H
+
+unsigned nth(unsigned);
+
+#endif