+++ Exercism/c/phone-number/src/phone_number.h
@@ -1,0 +1,11 @@
+#ifndef PHONE_NUMBER_H
+#define PHONE_NUMBER_H
+
+char* phone_number_clean(const char*);
+
+char* phone_number_get_area_code(const char*);
+
+char* phone_number_format(const char*);
+
+
+#endif